add UI

+++ src/Domain/Domain.Dtos/NumberIntoWord.cs
@@ -1,9 +1,11 @@
+using System;
+
 namespace Domain.Dtos
 {
     public class NumberIntoWord
     {
         public string Name { get; set; }
-        public int InputNumber { get; set; }
+        public long InputNumber { get; set; }
         public string NumberIntoWords { get; set; }
     } 
 }
