+++ src/Domain/Domain.Interface/Infrastructure/Providers/INumberIntoWordsProvider.cs
@@ -1,9 +1,10 @@
 using Domain.Dtos;
+using System;
 
 namespace Domain.Interfaces.Infrastructure.Providers
 {
     public interface INumberIntoWordsProvider
     {
-        ApiReturn<NumberIntoWord> ReturnNumberIntoWords(string name, int inputNumber);
+        ApiReturn<NumberIntoWord> ReturnNumberIntoWords(string name, long inputNumber);
     } 
 }
